+++ scriptabit/plugins/csv_tasks.py
@@ -26,7 +26,6 @@ class CsvTasks(scriptabit.IPlugin):
         super().__init__()
         self.tasks = []
         self.tag_names = []
-        self.__print_help = None
 
     @staticmethod
     def supports_dry_runs():
@@ -55,7 +54,7 @@ class CsvTasks(scriptabit.IPlugin):
             metavar='FILE',
             help='CSV file for bulk task import')
 
-        self.__print_help = parser.print_help
+        self.print_help = parser.print_help
 
         return parser
 
@@ -94,7 +93,7 @@ class CsvTasks(scriptabit.IPlugin):
         """
         if not self._config.csv_file:
             logging.getLogger(__name__).warning('No CSV file specified')
-            self.__print_help()
+            self.print_help()
             return False
 
         logging.getLogger(__name__).info(